+++ lib/teamtailor/parser.rb
@@ -15,6 +15,7 @@ require "teamtailor/parser/referral"
 require "teamtailor/parser/partner_result"
 require "teamtailor/parser/requisition"
 require "teamtailor/parser/requisition_step_verdict"
+require "teamtailor/parser/upload"
 
 module Teamtailor
   class Parser
@@ -41,6 +42,7 @@ module Teamtailor
         when "partner-results" then Teamtailor::PartnerResult.new(record, included)
         when "requisitions" then Teamtailor::Requisition.new(record, included)
         when "requisition-step-verdicts" then Teamtailor::RequisitionStepVerdict.new(record, included)
+        when "uploads" then Teamtailor::Upload.new(record, included)
 
         else
           raise Teamtailor::UnknownResponseTypeError, record&.dig("type")