+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Unreleased
+- Added `Teamtailor::Upload` and `Client#create_upload`
 
 ## v0.3.3 - 2021-03-10